python packaging integration
Added python packing integration for boundlessgeo-schema. All python files are in the `python` directory. To package the distribution you will need to run `sh ./script/python.sh`. This will build a distribution in the `python/dist` directory. Steps have been added to the schema README.md and SPATIALIO_IMPORT has been added to the actions.json.

If doing a pip install from the github endpoint the `actions.p` file will need to be committed after running the python.sh script, since the script converts the actions.json to a actions.p (pickle) for the python package to import as a dictionary.